Fix mongodb crash if project num isn't found
The crash is not reproducible for the other frameworks as those patches
do not persist on long lived objects and thus execute after hooks have
been disabled. This PR also fixes a race condition crash that could
occur if hooks were disabled between the start and end of a single span
(resulting in a null dereference of the agent).

Fixes #149