Update the use of `sed` for system independence
`sed` was being used in the `http-e2e.js` test with the `-i` flag
to perform changes in-place.  However, this option has a different
syntax on OS X vs Linux.  As a fix, `-i.bak` is now used to
explicitly specify a backup extension works on both OS X and
Linux.

The command has been tested on OS X and Debian.