`startAgent()` is no longer idempotent
In particular, the `stop()` method of the agent returned by the
`startAgent()` method must be called before the `startAgent()`
method can be called again.  Otherwise, calling `startAgent()`
on an already running agent will throw an error.

Most of the unit tests pass, but work still needs to be done
on a couple remaining tests.