The module now exports a constructor
Now this module exports a constructor.  In addition, the `start`
method has been renamed to `startAgent`, and the agent must be
stopped before it can be started again.

These changes were made to more closely align this module with
the other Google Cloud API Client Libraries via the
`google-cloud` modules in [1].

[1]: https://github.com/GoogleCloudPlatform/google-cloud-node