context from childspan and other misc. changes
After some discussion with Matt & Kelvin, it was indeed correct to get
the context from the child span. Also, createChildSpan now has the
responsibility of checking for context loss, so we don't even need to
check for nullness on the rootSpan – instead the childSpan would be
null. This means that there is no need to get the root span in
child-only plugins.

Made some other cosmetic changes. api.constants is now available so
we don't need to require('../constants') anymore.