Ensure http response streams are paused
Responses returned by http#request are yielded in paused mode. By
attaching a 'data' listener to the request the http instrumentation was
switching the stream to flowing mode which could cause the request to
drain before the calling framework had a chance to attach their own
listeners.

This also removes a test for tracking span duration for http requests
with responses that were never checked. We are actually not able to
properly close spans in this case without changing the behavior of the
requests so this test should not pass.

Fixes #437