Initial commit
This is still a WIP, but is a starting point for having the trace
agent's tests run in an isolated way.

The goal of this is so that if a developer has docker and
docker-compose installed, they can just git clone
cloud-trace-nodejs and run `./docker/docker-run-tests.sh` to run
the tests.