Update deps, drop support for 0.12
This change updates @google-cloud/common to a version which is no longer
compatable with node v0.12. To upgrade this dependency, we must drop
support for 0.12 as well. Additionally, we will stop running some of the
non-interference tests which only ran on 0.12. These tests will be
upgraded to work with newer versions of node in follow on work.