chore: copy cached packages in appveyor (#642)
PR-URL: #642

The wildcard files that are `npm pack`'d into the package published to npm that are in `build` are all also in `build/src`, so remove `build/src` instead of `build`. This allows us to take advantage of the AppVeyor cache in Node 8 to store test plugin fixtures, because AppVeyor can only restore the cache at the beginning of a job and save it at the end. (If this change wasn't here, `npm prepare`, which is run by `npm install` and `npm run check-install`, would delete these fixtures, costing time to re-initialize them.)