chore: disable strictBindCallApply TSC rule
Use of the `arguments` object is prolific because we want to
non-intrusively monkeypatch functions, as it ensures that the arguments
passed to a user function are the same arguments that are passed through
this library's monkeypatches. For this reason the benefits of replacing
arguments with TS-preferred `args` as encouraged by the
strictBindCallApply rule are outweighed by the drawbacks.